lulu; add sequences back to the output table, and edit process logging

+++ src/pipecraft-core/service_scripts/lulu.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Post-clustering of OTU/ASV table with LULU 
-# Input = OTU table and correspoding fasta file
+# Input = OTU table (2nc col may be 'Sequence') and correspoding fasta file
 
 ################################################
 ###Third-party applications:
@@ -18,6 +18,13 @@
     #https://github.com/torognes/vsearch
 #pigz
 ##############################################
+
+# Checking tool versions
+vsearch_version=$(vsearch --version 2>&1 | head -n 1 | awk '{print $2}' | sed -e "s/,//g")
+blast_version=$(blastn -version 2>&1 | head -n 1 | awk '{print $2}')
+printf "# Checking tool versions ...\n"
+printf "# vsearch (version $vsearch_version)\n"
+printf "# BLAST (version $blast_version)\n"
 
 #env variables
 workingDir=${workingDir}
@@ -48,12 +55,14 @@ mkdir -p $output_dir
 regex='[^/]*$'
 otu_table_temp=$(echo $table | grep -oP "$regex")
 otu_table=$(printf "/extraFiles/$otu_table_temp")
+export otu_table # for lulu.R
 printf "\n input table = $otu_table \n"
 
 #get specified input fasta file
 regex='[^/]*$'
 input_fasta_temp=$(echo $fasta_file | grep -oP "$regex")
 input_fasta=$(printf "/extraFiles2/$input_fasta_temp")
+export input_fasta # for lulu.R
 printf "\n input fasta = $input_fasta \n"
 
 #############################
@@ -116,21 +125,30 @@ else
     end_process
 fi
 
-#copy OTU table to $output_dir for lulu
-cp $otu_table $output_dir/OTU_tab_for_lulu.txt
-
 ###Run LULU in R
 printf "# Running lulu\n"
 errormessage=$(Rscript /scripts/submodules/lulu.R 2>&1)
-echo $errormessage > $output_dir/R_run.log 
+echo $errormessage > $output_dir/lulu_R_run.log 
 wait
 printf "\n LULU completed \n"
+
+# format R log
+sed -i 's/;; /\n/g' $output_dir/lulu_R_run.log
+sed -i '/progress:/d' $output_dir/lulu_R_run.log
+
+
+
+# get output table name 
+lulu_table=$(basename $otu_table | sed 's/\.[^.]*$//')".lulu.txt"
+
+# get output fasta name 
+lulu_fasta=$(basename $input_fasta | sed 's/\.[^.]*$//')".lulu.fasta"
 
 # Generate new OTUs.fasta file that excluded "discarded" OTUs by lulu
 checkerror=$(seqkit grep --quiet \
                          --line-width 0 \
                          -f $output_dir/lulu_out_OTUids.txt \
-                         $input_fasta > $output_dir/lulu_out_RepSeqs.fasta 2>&1)
+                         $input_fasta > $output_dir/$lulu_fasta 2>&1)
 check_app_error
 
 ########################################
@@ -140,11 +158,13 @@ for i in $input_fasta.n*; do
     [[ -f $i ]] || continue
     rm -f "$i"
 done
-if [[ -f $output_dir/R_run.log ]]; then
-    rm -f $output_dir/R_run.log
-fi
-if [[ -f $output_dir/lulu_out_OTUids.txt ]]; then
-    rm -f $output_dir/lulu_out_OTUids.txt
+if [[ $debugger != "true" ]]; then
+    # if [[ -f $output_dir/lulu_R_run.log ]]; then
+    #     rm -f $output_dir/lulu_R_run.log
+    # fi
+    if [[ -f $output_dir/lulu_out_OTUids.txt ]]; then
+        rm -f $output_dir/lulu_out_OTUids.txt
+    fi
 fi
 
 end=$(date +%s)
@@ -159,27 +179,30 @@ if [[ $match_list_soft == "vsearch" ]]; then
     match_list_generation=$"vsearch --usearch_global $input_fasta --db $input_fasta --strand both --self --id $vsearch_perc_identity --iddef $vsearch_similarity_type --userout match_list.lulu --userfields query+target+id --maxaccepts 0 --query_cov $vsearch_coverage_perc"
 fi
 #count merged units and curated units
-curated_units=$(grep -c "^>" $output_dir/lulu_out_RepSeqs.fasta)
+curated_units=$(grep -c "^>" $output_dir/$lulu_fasta)
 merged_units=$(wc -l $output_dir/discarded_units.lulu | awk '{print $1}')
 if (( $merged_units == 0 )); then
     info=$"No output table generated; OTU/ASV counts remained the same with the selected settings."
     rm $output_dir/discarded_units.lulu
-    rm $output_dir/lulu_out_RepSeqs.fasta
-    rm $output_dir/lulu_out_table.txt
-    rm $output_dir/lulu_out_table.csv
+    rm $output_dir/$lulu_fasta
+    rm $output_dir/$lulu_table
 else
     info=$"Output table consists of $curated_units Features (OTUs/ASVs)."
 fi
 
 printf "# Performed post-clustering with LULU (see 'Core commands' below for the used settings).
 
+Start time: $start_time
+End time: $(date)
+Runtime: $runtime seconds
+
 Total of $merged_units Features (OTUs/ASVs) were merged.
 $info
 
 Files in 'lulu_out' directory:
 ------------------------------
-# lulu_out_table.txt     = curated table in tab delimited txt format
-# lulu_out_RepSeqs.fasta = fasta file for the Features (OTUs/ASVs) in the curated table
+# $lulu_table     = curated table in tab delimited txt format
+# $lulu_fasta     = fasta file for the Features (OTUs/ASVs) in the curated table
 # match_list.lulu        = match list file that was used by LULU to merge 'daughter molecular' units
 # discarded_units.lulu   = Features (OTUs/ASVs) that were merged with other units based on specified thresholds)
 
@@ -193,9 +216,9 @@ Total run time was $runtime sec.
 ###Third-party applications:
 #lulu v0.1.0
     #citation: Froslev, T.G., Kjoller, R., Bruun, H.H. et al. 2017. Algorithm for post-clustering curation of DNA amplicon data yields reliable biodiversity estimates. Nat Commun 8, 1188.
-#BLAST 2.14.0+ (if BLAST was used to make match_list.lulu)
+#BLAST (version $blast_version) (if BLAST was used to make match_list.lulu)
     #citation: Camacho C., Coulouris G., Avagyan V., Ma N., Papadopoulos J., Bealer K., & Madden T.L. (2008) BLAST+: architecture and applications. BMC Bioinformatics 10:421. 
-#vsearch v2.23.0 (if vsearch was used to make match_list.lulu)
+#vsearch (version $vsearch_version) (if vsearch was used to make match_list.lulu)
     #citation: Rognes T, Flouri T, Nichols B, Quince C, Mahé F (2016) VSEARCH: a versatile open source tool for metagenomics PeerJ 4:e2584
 ##############################################" > $output_dir/README.txt
